Inline the smallest images, saving 3 requests and ~70kB
Frank and Fernanda Jumping.png are inlined, favicon.ico was converted to PNG (256x256) (from 64kB down to 400 bytes) and then also inlined.

It is possible (and would improve performance) to inline the other images too, but would make the html more unreadable.
Since there is no php the files can't be included on request, but I assume it would be possible to modify the deploy action or whatever to inline the media at that time.
Maybe @Tr1NgleDev knows how to do that. PNG basically needs a base64 convert of the data, then you prepend `data:image/png;base64,` to that and replace the url with the resulting string.

Candidates for inlining are: The forg GIFs, the logo, and the TTF. Only the video has reason to be a dedicated file.